Hook up new ast-based input-finding for python apps and also make input-finding interfaces the same across both languages
